Fit progress handling is moved to FitSuiteManager.
